Allow a custom markdown identifier for all types.
This commit adds a config parameter markdown_identifier which
allows changing the identifier this plugin looks for when parsing
a markdown file.

Since the images did not support this kind of syntax it was
introduces as well.

```
Figure: Captions {<figure properties>}

![](img.jpg)
```

This change was necessary to allow adding attributes to the
figure element directly. Previously this was only possible
for the img elements directly.