Add Unittest
This commit adds unittest to all the Identifiers and tests all
their configurations. It also tests if the plugin works with
MkDocs by building a demo site that contains all Identifiers.

These tests give a better feedback if changes break somthing
or change the exitisting behaviour. They are be all means not
complete but a start.