Allow markdown syntax within captions (#14)
This commit changes the logic of the preprocessing that wraps the
table caption such that the caption is self is still plain markdown.
This causes the mkdocs parser to convert the caption into html by
its own, which allows the usage of markown in captions.

The logic is adapted in such a way that a custom html element is
created both befor and after the caption. The caption itself is
untouched. In the post processing the two added tags can be used
to find the start and end of the caption easily.